Create script.py
This is the py file to clone from. Keep in mind this is not perfect whatsoever as I do not have much experience in programming and so far I've tried it for a few cases and it has worked. I must implement many other things before being actually completed. I hope it helps you tho. REMEMBER TO RUN AS ADMIN FROM CONSOLE!!!!